Refactor : footer, header 슬라이스 이동
