+++ src/widgets/index.ts
@@ -1,2 +1,1 @@
-export { Header } from "./Header"
-export { Footer } from "./Footer"
+export { Header, Footer } from "./layout"
