+++ src/widgets/layout/Footer.tsx
